requirements.txt
pyhive
requests
python-dotenv
pandas
ipywidgets